+++ src/pyunrealsdk/unreal_bindings/bindings.cpp
@@ -1,10 +1,12 @@
 #include "pyunrealsdk/pch.h"
 #include "pyunrealsdk/unreal_bindings/bindings.h"
 #include "pyunrealsdk/unreal_bindings/bound_function.h"
+#include "pyunrealsdk/unreal_bindings/persistent_object_ptr_property.h"
 #include "pyunrealsdk/unreal_bindings/property_access.h"
 #include "pyunrealsdk/unreal_bindings/uenum.h"
 #include "pyunrealsdk/unreal_bindings/uobject.h"
 #include "pyunrealsdk/unreal_bindings/uobject_children.h"
+#include "pyunrealsdk/unreal_bindings/weak_pointer.h"
 #include "pyunrealsdk/unreal_bindings/wrapped_array.h"
 #include "pyunrealsdk/unreal_bindings/wrapped_struct.h"
 #include "unrealsdk/unreal/classes/ufield.h"
@@ -28,6 +30,8 @@ void register_module(py::module_& mod) {
     register_wrapped_array(unreal);
     register_wrapped_struct(unreal);
     register_bound_function(unreal);
+    register_weak_pointer(unreal);
+    register_persistent_object_properties(unreal);
 }
 
 }  // namespace pyunrealsdk::unreal